% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_Functions.R
\name{bal.value}
\alias{bal.value}
\title{The balance value for a given matrix of counts using the variables appearing
in a \code{data.frame} like the output of \code{selbal.aux}}
\usage{
bal.value(bal.tab, x)
}
\arguments{
\item{bal.tab}{a \code{data.frame} including the variables defining the
balance (like the output of \code{selbal.aux} or the sixth element in the
output of \code{selbal.cv} function).}

\item{x}{the matrix with the log-transformed counts for a given subset of
individuals.}
}
\value{
A \code{vector} with the balance values for each subject.
}
\description{
The balance value for a given matrix of counts using the variables appearing
in a \code{data.frame} like the output of \code{selbal.aux}
}
\examples{
# Load data set
  load("HIV.rda")
# Define x and y
  x <- HIV[,1:60]
  y <- HIV[,62]
# Run the algorithm
  Bal <- selbal.aux(x,y)
# Balance values for the individuals
  bal.value(Bal,x)

}
