% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selbal_Functions.R
\name{cmultRepl2}
\alias{cmultRepl2}
\title{Zero replacement for compositional data}
\usage{
cmultRepl2(x, zero.rep = "bayes")
}
\arguments{
\item{x}{a \code{matrix} object with the information of variables
(\emph{columns}) for each sample (\emph{rows}).}

\item{zero.rep}{if \emph{"bayes"} the Bayesian - Multiplicative treatment
implemented in \code{zCompositions} is applied. If \emph{"one"}, a
pseudocount of 1 is added to the whole matrix.}
}
\value{
The initial matrix after the zero - replacement normalized so that
each sample's composition sums one.
}
\description{
\code{cmultRepl2} replaces the zeros for a matrix where each row is
compositional
}
\examples{

# Load the count matrix (with zeros)
  x <- HIV[,1:60]
# Zero replacement
  x.non0 <- cmultRepl2(x, zero.rep = "bayes")


}
