% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Crohn}
\alias{Crohn}
\title{Microbiome composition related to Crohn`s disease study}
\format{The \code{data.frame} is composed by 48 genera and a variable
\describe{
  \item{genera}{The first 48 columns, from \emph{g_Turicibacter} until
  \emph{g_Bilophila} referred to different genera.}
  \item{y}{a \code{factor} indicating if the sample corresponds to a case (
  \emph{CD}) or a control (\emph{no}).}
}}
\description{
A dataset containing the number of counts of 48 different genera in a group
of 975 samples (including 662 samples of patients with Crohn`s disease and
313 controls).
The \code{data.frame} is composed by 48 genera and a factor variable
}
\references{
\url{https://qiita.ucsd.edu/study/description/1939}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
