% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HIV}
\alias{HIV}
\title{Microbiome, HIV infection and MSM factor}
\format{The \code{data.frame} is composed by 60 genera and 2 variables
\describe{
  \item{genera}{The first 60 columns, from \emph{g_Prevotella} until
       \emph{o_NB1-n_g_unclassified} referred to different genera.}
  \item{MSM}{a factor determining if the individual is \code{MSM} (\emph{Men Sex with
   Men}) or not (\code{nonMSM}).}
  \item{HIV_Status}{a factor specifying if the individual is infected
   (\code{Pos}) or not (\code{Neg}).}

}}
\description{
A dataset containing the number of counts of 60 different genera in a group
of 155 samples (including HIV - infected and non - infected patients).
The \code{data.frame} is composed by 60 genera and two variables.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/27077120}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
