% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_Functions.R
\name{logit.cor}
\alias{logit.cor}
\title{Computes an association value  between a dichotomous variable and a
continuous one.}
\usage{
logit.cor(FIT, y, logit.acc)
}
\arguments{
\item{FIT}{a \code{glm} object referred to the logistic regression of a
dichotomous variable.}

\item{y}{the response variable (dichotomous).}

\item{logit.acc}{when \code{y} is dichotomous, the measure to compute for
the correlation between \code{y} and the proposed \emph{balance}
adjusting for covariates. One of the following values: \code{"Rsq"} (default),
 \code{"AUC"} or \code{"Tjur"}.}
}
\value{
The association value using the selected method \code{logit.acc}.
}
\description{
Computes an association value  between a dichotomous variable and a
continuous one.
}
