% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selbal_Functions.R
\name{plot.bal}
\alias{plot.bal}
\title{The balance value for a given matrix of counts using the variables appearing
in a data.frame like the output of selbal.aux}
\usage{
\method{plot}{bal}(POS, NEG, x, y, covar = NULL, col, logit.acc = NULL)
}
\arguments{
\item{POS}{a vector containing the variables taking part in the numerator of
the balance.}

\item{NEG}{a vector containing the variables taking part in the denominator
of the balance.}

\item{x}{a \code{matrix} object with the log-transforemd counts}

\item{y}{the response variable, either continuous or dichotomous.}

\item{covar}{a \code{data.frame} with the covariates to adjust for.}

\item{col}{a vector with two values for representing the individuals for
each of the two considered groups.}

\item{logit.acc}{when \code{y} is dichotomous, the measure to compute for
the correlation between \code{y} and the proposed \emph{balance}
adjusting for covariates. One of the following values: \code{"AUC"} (default),
\code{"Dev"}, \code{"Rsq"} or \code{"Tjur"}.}
}
\value{
A graphical representation of the balance for the selected
individuals and the variables taking part on it.
}
\description{
The balance value for a given matrix of counts using the variables appearing
in a data.frame like the output of selbal.aux
}
