% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selbal_Functions.R
\name{plot.tab}
\alias{plot.tab}
\title{Plots the cross - validated summary table}
\usage{
\method{plot}{tab}(dat, col = c("steelblue1", "tomato1"))
}
\arguments{
\item{dat}{the \code{data.frame} object to draw obtained from
\code{selbal.cv} function.}

\item{col}{vector with two colors, each of them referring to the variables
appearing in the numerator and in the denominator, respectively.}
}
\value{
A colored table with the information of the given \code{data.frame}:

\itemize{
 \item The first column of the table (%) represents the proportion of times
 a variables has been selected for a balance.

 \item The second column (\emph{Global}) shows the result for the balance
 obtained using all the available samples.

 \item The last three columns represent the most repeated balances in the
 cross - validation procedure.

 }
}
\description{
Plots the cross - validated summary table
}
\examples{
#---------------------------------------------------#
# 1) Compute a cross - validated balance selection
#---------------------------------------------------#
 # Load data set
   load("HIV.rda")
  # Define x and y
    x <- HIV[,1:60]
    y <- HIV[,62]
 # Run the algorithm
    CV.Bal <- selbal.cv(x,y)

#----------------------------------------#
# 2) Plot the table
#----------------------------------------#

  plot.tab(CV.Bal[[4]])

}
