% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_Functions.R
\name{rename_OTU}
\alias{rename_OTU}
\title{Rename each taxon}
\usage{
rename_OTU(phy, rank, db = "SILVA")
}
\arguments{
\item{phy}{the phyloseq object where the information is contained.}

\item{rank}{a \code{numeric} value indicating the taxonomic level where the
names should be taken. It corresponds with the column number of
\code{phy@tax_table} associated to the desired taxonomic rank.}

\item{db}{the bacterial database used to name each OTU of the phyloseq
object: SILVA (\emph{default}) or GreenGenes.}
}
\value{
A vector with the names for each bacteria. It has not repeated names
and none of them starts with \emph{Incertae_Sedis} or \emph{unclassified}.
}
\description{
\code{rename_OTU} assigns a non - repeated name to each bacteria for a given
 taxonomic level.
}
