% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_Functions.R
\name{rowM}
\alias{rowM}
\title{Calculates the mean of each row of a matrix (though having only one column).}
\usage{
rowM(x)
}
\arguments{
\item{x}{a \code{matrix} object.}
}
\value{
A \code{vector} with the mean of each row of \code{x}, even if the
matrix only contains one column.
}
\description{
Calculates the mean of each row of a matrix (though having only one column).
}
\examples{
# Build a matrix with one column
  M <- matrix(rnorm(10), nrow=1)
# rowM (resulting on the same matrix M)
  rowM(M)

# Build a matrix
  M <- matrix (runif(100), nrow=10)
# Apply rowM function
  rowM(M)

}
