% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sCD14}
\alias{sCD14}
\title{Microbiome and sCD14 inflammation parameter}
\format{The \code{data.frame} is composed by 60 genera and a variable
\describe{
  \item{genera}{The first 60 columns, from \emph{g_Prevotella} until
  \emph{o_NB1-n_g_unclassified} referred to different genera.}
  \item{sCD14}{a \code{numeric} variable with the value of the inflammation
  parameter sCD14 for each sample.}
}}
\description{
A dataset containing the number of counts of 60 different genera in a group
of 151 samples (including HIV - infected and non - infected patients).
The \code{data.frame} is composed by 60 genera and a numeric variable
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
