% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_Functions.R
\name{selbal.aux}
\alias{selbal.aux}
\title{The balance defined in the order its variables were selected}
\usage{
selbal.aux(x, y, th.imp = 0, covar = NULL, logit.acc = "AUC", logt = T,
  maxV = 1e+10, zero.rep = "bayes")
}
\arguments{
\item{x}{a \code{matrix} object with the information of variables
(\emph{columns}) for each sample (\emph{rows}).}

\item{y}{the response variable, either continuous or dichotomous.}

\item{th.imp}{the minimum increment needed when adding a new variable into
the balance, in order to consider an improvement.}

\item{covar}{\code{data.frame} with the variables to adjust for
(\emph{columns}).}

\item{logit.acc}{when \code{y} is dichotomous, the measure to compute for
the correlation between \code{y} and the proposed \emph{balance}
adjusting for covariates. One of the following values: \code{"Rsq"} (default),
 \code{"AUC"} or \code{"Tjur"}.}

\item{logt}{a logical value indicating if the data needs to be
log-transformed.}

\item{maxV}{the number maximum of variables defining the balance.}
}
\value{
A \code{data.frame} with the following objects:

\itemize{
\item \code{Taxa} the name of the taxa added into the balance.
\item \code{Group} \emph{NUM} if the variable is added into the numerator
and \emph{DEN} if it is added in the denominator.

}
}
\description{
The balance defined in the order its variables were selected
}
\examples{
# Load data set
  load("HIV.rda")
# Define x and y
  x <- HIV[,1:60]
  y <- HIV[,62]
# Run the algorithm
  Bal <- selbal.aux(x,y)
}
