% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKG_Functions.R
\name{selbal}
\alias{selbal}
\title{Looks for a highly-associated balance with a response variable}
\usage{
selbal(x, y, th.imp = 0, covar = NULL, logit.acc = "AUC", logt = T,
  col = c("steelblue1", "tomato1"), tab = T, draw = T, maxV = 1e+10,
  zero.rep = "bayes")
}
\arguments{
\item{x}{a \code{matrix} object with the information of variables
(\emph{columns}) for each sample (\emph{rows}).}

\item{y}{the response variable, either continuous or dichotomous.}

\item{th.imp}{a numeric value indicating the minimum increment required in
the association parameter between two consecutive steps in order to continue
with the variable addition into the \emph{balance}.}

\item{covar}{\code{data.frame} with the variables to adjust for
(\emph{columns}).}

\item{logit.acc}{when \code{y} is dichotomous, the measure to compute for
the correlation between \code{y} and the proposed \emph{balance}
adjusting for covariates. One of the following values: \code{"Rsq"} (default),
 \code{"AUC"} or \code{"Tjur"}.}

\item{logt}{\code{logical} value determining if \code{x} needs a
log-transformation. TRUE if \code{x} contains raw counts or proportions.}

\item{col}{\code{vector} of two colours for differentiate the variables
appearing in the numerator and in the denominator of the balances.}

\item{tab}{\code{logical} value. It specifies if a table with the variables
included in the balance (ordered) and the evolution of the association
parameter is demanded.}

\item{draw}{\code{logical} value to concretif a plot with the balance value
and the response variable is desired.}

\item{maxV}{\code{numeric} value defining the maximum number of variables
composing the balance. Default 1e10 to give prevalence to \code{th.imp}
parameter.}

\item{zero.rep}{a value defining the method to use for zero - replacement.
\code{"bayes"} for BM-replacement or \code{"one"} to add one read tho each
cell of the matrix.}

\item{opt.cri}{parameter for selecting the method to determine the optimal
number of variables. \code{"max"} to define this number as the number of
variables which maximizes the association value or \code{"1se"} to take also
the standard error into account.}
}
\value{
A \code{list} with the following objects:


\itemize{

\item \code{FINAL.BAL} the numeric values of the selected balance for each
sample.
\item \code{POS} a vector with the variables appearing in the numerator of
the \emph{balance}.
\item \code{NEG} a vector with the variables appearing in the denominator of
the \emph{balance}.
\item \code{INC.VAR} a vector with both \code{POS} and \code{NUM} variables
(included variables).
\item \code{ACC.Bal} a vector with the association value after each step of
the algorithm.
\item \code{EVOL} a \code{data.frame} with the variables sorted as they have
been added into the balance with the corresponding association value after
their inclusion. Only returned if \code{tab} is \code{TRUE}.
\item \code{FINAL.P} the graphical representation of the results. Only
showed if \code{draw = T}.
\item \code{FIT.Final} the regression model taking covariates and the final
balance as the explanatory variables and \code{y} as the response variable.

}
}
\description{
Looks for a highly-associated balance with a response variable
}
\examples{
# Load data set
  load("HIV.rda")
# Define x and y
  x <- HIV[,1:60]
  y <- HIV[,62]
# Run the algorithm
  Bal <- selbal(x,y)
}
